+++ app/src/main/java/org/hyperledger/besu/controller/BesuControllerBuilder.java
@@ -92,9 +92,7 @@ import org.hyperledger.besu.ethereum.trie.pathbased.bonsai.cache.BonsaiCachedMer
 import org.hyperledger.besu.ethereum.trie.pathbased.bonsai.cache.CodeCache;
 import org.hyperledger.besu.ethereum.trie.pathbased.bonsai.storage.BonsaiWorldStateKeyValueStorage;
 import org.hyperledger.besu.ethereum.trie.pathbased.bonsai.storage.flat.BonsaiArchiveFlatDbStrategy;
-import org.hyperledger.besu.ethereum.trie.pathbased.bonsai.worldview.BonsaiArchiver;
 import org.hyperledger.besu.ethereum.trie.pathbased.bonsaiarchive.BonsaiFlatDbToArchiveMigrator;
-import org.hyperledger.besu.ethereum.trie.pathbased.common.storage.PathBasedWorldStateKeyValueStorage;
 import org.hyperledger.besu.ethereum.trie.pathbased.common.storage.flat.CodeHashCodeStorageStrategy;
 import org.hyperledger.besu.ethereum.trie.pathbased.common.trielog.TrieLogManager;
 import org.hyperledger.besu.ethereum.trie.pathbased.common.trielog.TrieLogPruner;
@@ -898,15 +896,6 @@ public abstract class BesuControllerBuilder implements MiningConfigurationOverri
 
     if (DataStorageFormat.X_BONSAI_ARCHIVE.equals(
         dataStorageConfiguration.getDataStorageFormat())) {
-      final BonsaiWorldStateKeyValueStorage worldStateKeyValueStorage =
-          worldStateStorageCoordinator.getStrategy(BonsaiWorldStateKeyValueStorage.class);
-      final BonsaiArchiver archiver =
-          createBonsaiArchiver(
-              worldStateKeyValueStorage,
-              blockchain,
-              scheduler,
-              ((BonsaiWorldStateProvider) worldStateArchive).getTrieLogManager());
-
       if (worldStateStorageCoordinator.isMatchingFlatMode(FlatDbMode.FULL)
           || worldStateStorageCoordinator.isMatchingFlatMode(FlatDbMode.PARTIAL)) {
         final BonsaiFlatDbToArchiveMigrator archiveMigrator =
@@ -926,7 +915,6 @@ public abstract class BesuControllerBuilder implements MiningConfigurationOverri
                     LOG.info("Node is in sync, starting Bonsai archive migration");
                     archiveMigrator
                         .migrate()
-                        .thenRun(() -> blockchain.observeBlockAdded(archiver))
                         .exceptionally(
                             error -> {
                               LOG.error(
@@ -946,7 +934,6 @@ public abstract class BesuControllerBuilder implements MiningConfigurationOverri
         archiveMigrator.startOngoingMigration();
         // Close the migrator before storageProvider so callback finishes before RocksDB is closed
         closeables.addFirst(archiveMigrator);
-        blockchain.observeBlockAdded(archiver);
       }
     }
 
@@ -1036,24 +1023,6 @@ public abstract class BesuControllerBuilder implements MiningConfigurationOverri
     trieLogPruner.initialize();
 
     return trieLogPruner;
-  }
-
-  private BonsaiArchiver createBonsaiArchiver(
-      final WorldStateKeyValueStorage worldStateStorage,
-      final Blockchain blockchain,
-      final EthScheduler scheduler,
-      final TrieLogManager trieLogManager) {
-    final BonsaiArchiver archiver =
-        new BonsaiArchiver(
-            (PathBasedWorldStateKeyValueStorage) worldStateStorage,
-            blockchain,
-            scheduler::executeServiceTask,
-            trieLogManager,
-            metricsSystem);
-
-    archiver.initialize();
-    LOG.info("Bonsai archiver initialised");
-    return archiver;
   }
 
   private BonsaiFlatDbToArchiveMigrator createArchiveMigrator(