+++ ethereum/core/src/main/java/org/hyperledger/besu/ethereum/storage/keyvalue/KeyValueSegmentIdentifier.java
@@ -53,20 +53,6 @@ public enum KeyValueSegmentIdentifier implements SegmentIdentifier {
       false,
       true,
       true),
-  ACCOUNT_INFO_STATE_FREEZER(
-      "ACCOUNT_INFO_STATE_FREEZER".getBytes(StandardCharsets.UTF_8),
-      EnumSet.of(X_BONSAI_ARCHIVE),
-      true,
-      false,
-      true,
-      true),
-  ACCOUNT_STORAGE_FREEZER(
-      "ACCOUNT_STORAGE_FREEZER".getBytes(StandardCharsets.UTF_8),
-      EnumSet.of(X_BONSAI_ARCHIVE),
-      true,
-      false,
-      true,
-      true),
   VARIABLES(new byte[] {11}), // formerly GOQUORUM_PRIVATE_WORLD_STATE
 
   // previously supported GoQuorum private states